+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -3,7 +3,11 @@
 
 ### New Features
 - Update `DeviceCodeCredential` to output device code information and authentication instructions in the console, in the case no `deviceCodeCallback` is specified.
+  - Added `DeviceCodeCallback` to `DeviceCodeCredentialOptions`
+  - Added default constructor to `DeviceCodeCredential`
 
+### Breaking Changes
+- Replaced `DeviceCodeCredential` constructor overload taking `deviceCodeCallback` and `DeviceCodeCredentialOptions` with constructor taking only `DeviceCodeCredentialOptions`
 
 ## 1.3.0-beta.1 (2020-09-11)