+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 1.4.0-beta.1 (Unreleased)
+
+
 ## 1.3.0 (2020-11-12)
 
 ### New Features